+++ CHANGELOG.md
@@ -9,6 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Navigate to Recurring Transactions from Settings.
 
 
+## [0.29.0] - 2025-05-19
+### Added
+- View recurring transaction details with edit and delete actions.
+- View upcoming recurring transactions in a new screen.
+
 ## [0.28.0] - 2025-05-19
 ### Added
 - Set Budget now opens as a dialog similar to New Transaction.